backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
drf-spectacular==0.26.5
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
boto3==1.29.3
django-storages==1.14.2
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
whitenoise==6.6.0
gunicorn==21.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.1.0
factory-boy==3.3.0
pytest-django==4.7.0
coverage==7.3.2
